// packages/golden-audio-ui/generated/AudioDeviceDescriptor.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AudioDeviceFingerprint } from "./AudioDeviceFingerprint";
import type { AudioDeviceProfileKey } from "./AudioDeviceProfileKey";
import type { AudioDeviceTargetId } from "./AudioDeviceTargetId";
import type { PhysicalChannelDescriptor } from "./PhysicalChannelDescriptor";
import type { SupportedStreamConfiguration } from "./SupportedStreamConfiguration";

export type AudioDeviceDescriptor = { target: AudioDeviceTargetId, label: string, stable_id: boolean, fingerprint: AudioDeviceFingerprint, profile_key: AudioDeviceProfileKey, input_channels: Array<PhysicalChannelDescriptor>, output_channels: Array<PhysicalChannelDescriptor>, supported_configurations: Array<SupportedStreamConfiguration>, is_system_default_input: boolean, is_system_default_output: boolean, };